Rename WebviewBase to AbstractWebview
